Removed the dbCharcterService to run the queries directly from the service class.
