+++ package.json
@@ -53,7 +53,7 @@
     "@nestjs/jwt": "^6.0.0",
     "@nestjs/passport": "^6.0.0",
     "@nestjs/platform-fastify": "^6.0.2",
-    "@nestjs/swagger": "^3.0.1",
+    "@nestjs/swagger": "^3.0.2",
     "@nestjs/terminus": "^6.0.0",
     "bcryptjs": "^2.4.3",
     "class-transformer": "^0.2.0",
@@ -79,7 +79,7 @@
     "reflect-metadata": "^0.1.13",
     "rxjs": "^6.4.0",
     "tsconfig-paths": "^3.8.0",
-    "zone.js": "^0.8.29"
+    "zone.js": "^0.9.0"
   },
   "devDependencies": {
     "@abraham/reflection": "^0.5.1",
