+++ src/server/app/db/models/db_question.model.ts
@@ -1,4 +1,4 @@
 export class DbQuestion {
-  qId: string;
-  qQuestion: string;
+  dbId: string;
+  dbQuestion: string;
 }
