+++ src/server/app/zeldaplay/character/character.module.ts
@@ -4,12 +4,11 @@ import { CharacterController } from '@Character/character.controller';
 import { CharacterHealthIndicator } from '@Character/character.health';
 import { CharacterService } from '@Character/character.service';
 import { SharedServerModule } from '@Shared/shared.module';
-import { DbCharacterService } from './db-character/db-character.service';
 
 @Module({
   imports: [SharedServerModule],
   controllers: [CharacterController],
-  providers: [CharacterService, DbCharacterService, CharacterHealthIndicator],
+  providers: [CharacterService, CharacterHealthIndicator],
   exports: [CharacterHealthIndicator]
 })
 export class CharacterServerModule {}
