+++ src/server/swagger.ts
@@ -13,23 +13,26 @@ export function configSwagger() {
       'https://github.com/jmcdo29/zeldaPlay/blob/master/LICENSE'
     )
     .addTag(
-      'character',
+      'Character',
       'APIs that specifically pertain to the character or character saves'
     )
     .addTag(
-      'note',
+      'Note',
       'APIs for getting the Notes of and adding new Notes for the character'
     )
     .addTag(
-      'spell',
+      'Spell',
       'APIs for getting the Spells of and adding new Spells for the character'
     )
-    .addTag('skill', 'APIs for getting new Skills for the character')
+    .addTag('Skill', 'APIs for getting new Skills for the character')
     .addTag(
-      'weapon',
+      'Weapon',
       'APIs for getting the Spells of and adding new Weapons for the character'
     )
-    .addTag('user', 'APIs for registering and logging in to the application')
+    .addTag(
+      'Sign In/ Sign Up/ Sign Out',
+      'APIs for registering and logging in to the application'
+    )
     .addBearerAuth('Authorization', 'header')
     .build();
 }
